% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affineCipher.R
\name{affineCipher}
\alias{affineCipher}
\title{Affine Cipher}
\usage{
affineCipher(plainText, alpha, beta)
}
\arguments{
\item{plainText}{A string of lowercase letters}

\item{alpha}{An integer (mod 26) to scale by}

\item{beta}{An integer (mod 26) to shift by}
}
\value{
An encrypted (or decrypted) string, where each letter has been scaled by \code{alpha}, and shifted by \code{beta} places.
}
\description{
Encrypt a string using an affine cipher
}
\examples{
affineCipher("examplemessage", 9, 7)
}

