% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsaTools.r
\name{bigzToString}
\alias{bigzToString}
\title{Convert (some) big integers to strings}
\usage{
bigzToString(n)
}
\arguments{
\item{n}{A `bigz` integer, whose raw representation must not contain 00's}
}
\value{
A character string
}
\description{
This function is intended to serve as an inverse to the \code{\link{stringToBigz}} function.
If the raw representation of `n` contains 00's, this function will produce an
`embedded nul in string` error. Thus it is not suitable for all integers.
}
\examples{
stringAsInt <- stringToBigz("Any string.")
bigzToString(stringAsInt)
# bigzToString(256^3+256) # embedded nul error!
}

