% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeTools2.R
\name{coset}
\alias{coset}
\title{Coset of a code given a representative}
\usage{
coset(r, C)
}
\arguments{
\item{r}{An n-dimensional numeric vector of 0's and 1's.}

\item{C}{An [n,k] binary linear code, represented as a 2^k by n matrix. The rows of
the matrix are the code words.}
}
\value{
A 2^k by n matrix, whose rows are all the elements of the coset r + C
in the vector space of all binary vectors of dimension n.
}
\description{
Coset of a code given a representative
}
\examples{
coset(c(1,1,1,1), generateCode(matrix(c(1,0,1,1,
                                        0,1,1,0),
                                        nrow=2, byrow=TRUE))) # see page 412
}

