% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeTools2.R
\name{cosetLeader}
\alias{cosetLeader}
\title{Coset leader}
\usage{
cosetLeader(r, C)
}
\arguments{
\item{r}{An n-dimensional numeric vector of 0's and 1's.}

\item{C}{An [n,k] binary linear code, represented as a 2^k by n matrix. The rows of
the matrix are the code words.}
}
\value{
An n-dimensional vector, which is a vector of minimal weight in the
coset r + C.
}
\description{
Coset leader
}
\examples{
cosetLeader(c(1,1,1,1), generateCode(matrix(c(1,0,1,1,
                                        0,1,1,0),
                                        nrow=2, byrow=TRUE))) # see page 412
}

