% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeTools2.R
\name{cosetLeaderSyndromeTable}
\alias{cosetLeaderSyndromeTable}
\title{Coset leader syndrome table}
\usage{
cosetLeaderSyndromeTable(G)
}
\arguments{
\item{G}{A generator matrix for a linear binary code.}
}
\value{
A of all the coset leaders indexed by their corresponding syndromes.
}
\description{
Coset leader syndrome table
}
\examples{
cosetLeaderSyndromeTable(matrix(c(1,0,1,1,
                                  0,1,1,0),
                                  nrow=2, byrow=TRUE)) # see page 412
}

