% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigenereCipher.R
\name{devigenere}
\alias{devigenere}
\title{Decrypt Vigenere}
\usage{
devigenere(txt, keyVector)
}
\arguments{
\item{txt}{The text which you are decrypting}

\item{keyVector}{A numerical vector used as the encryption key}
}
\description{
Decrypt a message encoded with a vigenere cipher with a known key
}
\examples{
devigenere("moqkeaesftlvmpmsom", stringToMod26("thiswillbeavector"))
}

