% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigenereCipher.R
\name{digramTable}
\alias{digramTable}
\title{Build Digram Table}
\usage{
digramTable(txt)
}
\arguments{
\item{txt}{The text which you are analyzing}
}
\description{
Builds a table displaying thee frequencies of digrams
}
\examples{
digramTable("thisisan@examplesforbuildingadigramtablealongertextwouldbemoreuseful")
}

