% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_logs.R, R/test.R
\name{discreteLogBrute}
\alias{discreteLogBrute}
\title{Discrete Logarithm (brute force)}
\usage{
discreteLogBrute(alpha, beta, p)

discreteLogBrute(alpha, beta, p)
}
\arguments{
\item{alpha}{The base, assumed to be a primitive root modulo p.}

\item{beta}{An element of Z_p}

\item{p}{A prime modulus.}

\item{alpha}{The base, assumed to be a primitive root modulo p.}

\item{beta}{An element of Z_p}

\item{p}{A prime modulus.}
}
\value{
Computes L_alpha(beta), the smallest nonnegative integer x such that beta = alpha^x, modulo p.

Computes L_alpha(beta), the smallest nonnegative integer x such that beta = alpha^x, modulo p.
}
\description{
Computes a discrete logarithm by brute force.

Computes a discrete logarithm by brute force.
}
\examples{
discreteLogBrute(2,9,11) # should be 6
discreteLogBrute(2,9,11) # should be 6
}

