% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elipticalCurves.R
\name{ecNeg}
\alias{ecNeg}
\title{Elliptic curve additive inverse}
\usage{
ecNeg(p)
}
\arguments{
\item{p}{A length 2 vector representing a point on an elliptic curve.}
}
\value{
A length 2 vector representing the additive inverse of \code{p}.
}
\description{
Computes the additive inverse (negative) of a point on an elliptic curve. Note that
the output of this function does not depend on the choice of elliptic curve, or even
on the field.
}
\examples{
ecNeg(c(3,4)) # see page 351
}

