% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elipticalCurves.R
\name{ecPlotReal}
\alias{ecPlotReal}
\title{Plot a real elliptic curve}
\usage{
ecPlotReal(b, c, xmin = -5, xmax = 10, ymin = -30, ymax = 30, asp = 1)
}
\arguments{
\item{b}{A real number representing the coefficient b in the equation of the curve.}

\item{c}{A real number representing the coefficient c in the equation of the curve.}

\item{xmin}{Lower limit for x}

\item{xmax}{Upper limit for x}

\item{ymin}{Lower limit for y}

\item{ymax}{Upper limit for y}

\item{asp}{Aspect ratio}
}
\value{
A plot of the ellptic curve.
}
\description{
Creates a plot of the elliptic curve y^2 = x^3 + bx + c. Uses the \code{contour} plot
function.
}
\examples{
ecPlotReal(-5,1)
}

