% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elipticalCurves.R
\name{ecPowReal}
\alias{ecPowReal}
\title{Real elliptic curve "exponentiation"}
\usage{
ecPowReal(b, c, p, n)
}
\arguments{
\item{b}{A real number representing the coefficient b in the equation of the curve.}

\item{c}{A real number representing the coefficient c in the equation of the curve.}

\item{p}{A length 2 numeric vector, representing a point on the curve.}

\item{n}{An integer}
}
\value{
A length 2 vector representing \code{n} times \code{p} (on the elliptic curve).
}
\description{
Give a point \code{p} on a real elliptic curve y^2 = x^3 + bx + c and an integer \code{n},
returns the point np, that is, p added to itself n times. If n is negative, returns -p added
to itself n times. If n is zero, returns the point at NAinity. If p is the point at NAinity, 
also returns the point at NAinity. We call this operation "exponentiation", but since we 
represent elliptic curves with additive notation, "exponentiation" is repeated addition.
}
\examples{
ecPowReal(0, 73, c(-4, -3), 2) # See page 351
}

