% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsaTools.r
\name{fermatTest}
\alias{fermatTest}
\title{This tests to see if a given number (n) is a psedoprime for the given base (a).}
\usage{
fermatTest(n, a)
}
\arguments{
\item{n}{The number you are testing.}

\item{a}{The base for the Fermat Test}
}
\value{
TRUE, if the number is a psedoprime for the given base, else FALSE
}
\description{
This tests to see if a given number (n) is a psedoprime for the given base (a).
}
\examples{
fermatTest(bigPrime, 17)
fermatTest(bigNotPrime, 20)
fermatTest(bigPrime, bigNotPrime)
}

