% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigenereCipher.R
\name{findVigKey}
\alias{findVigKey}
\title{Find Vigenere Key}
\usage{
findVigKey(ciphertext, keyLength)
}
\arguments{
\item{ciphertext}{The encrypted text}

\item{keyLength}{The length of the ecryption key}
}
\description{
This finds the Vigenere key for a given ciphertext when the keylength is known.
}

