% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeTool.R
\name{generateCode}
\alias{generateCode}
\title{Generate a code from a basis}
\usage{
generateCode(B)
}
\arguments{
\item{B}{A list of basis code words, represented as a matrix of 0's
and 1's. The rows of this matrix are the basis code words. Assume that
the maximum size of a basis is 32. (The function does not require that
the rows of B be linearly independent.)}
}
\value{
A matrix consisting of all possible, distinct, linear combinations
of the basis code words over Z_2.
}
\description{
Generate a code from a basis
}
\examples{
generateCode(matrix(c(1,0,0,0,1,1,0,
                      0,1,0,0,1,0,1,
                      0,0,1,0,0,1,1,
                      0,0,0,1,1,1,1), nrow=4, byrow=TRUE))
}

