% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeTools2.R
\name{generatorMatrix}
\alias{generatorMatrix}
\title{Create generator matrix from parity check matrix}
\usage{
generatorMatrix(H)
}
\arguments{
\item{H}{A parity check matrix of the form [P^T I_{n-k}].}
}
\value{
A generator matrix of the form [I_k P].
}
\description{
Given a parity check matrix for a systematic linear binary code,
return the generator matrix.
}
\examples{
generatorMatrix(matrix(c(1,1,0,1,1,0,0,
                         1,0,1,1,0,1,0,
                         0,1,1,1,0,0,1), nrow=3, byrow=TRUE))
}

