% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeTool.R
\name{hammingDistance}
\alias{hammingDistance}
\title{Hamming distance}
\usage{
hammingDistance(c1, c2)
}
\arguments{
\item{c1}{A vector of 0's and 1's of length 2.}

\item{c2}{A vector of 0's and 1's of length 2.}
}
\value{
The number of positions where \code{c1} and \code{c2} differ.
}
\description{
Compute the Hamming distance between two n-dimensional vectors in Z_2.
}
\examples{
hammingDistance(c(1,0,0,1,1), c(0,1,0,1,0))
}

