% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTPcipher.R
\name{hillCipher}
\alias{hillCipher}
\title{Hill Cipher}
\usage{
hillCipher(txt, keyMatrix)
}
\arguments{
\item{txt}{the plaintext}

\item{keyMatrix}{the matrix containing the key used for encryption}
}
\value{
The encrypted text
}
\description{
Encode plaintext using the Hill cipher
}
\examples{
hillCipher("thisisanexample",matrix(c(1,3, 4,2, 9,5),nrow=3,ncol=2))

}

