% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeTool.R
\name{isPerfect}
\alias{isPerfect}
\title{Determine whether a code is perfect}
\usage{
isPerfect(C)
}
\arguments{
\item{C}{A code, represented as a matrix of 0's and 1's. The rows
of this matrix are the words of the code.}
}
\value{
TRUE if the code is perfect, FALSE if not.
}
\description{
Given a code C, determines whether the number of codewords equals the Hamming
sphere packing bound.
}
\examples{
isPerfect(generateCode(matrix(c(1,0,0,0,1,1,0,
                                0,1,0,0,1,0,1,
                                0,0,1,0,0,1,1,
                                0,0,0,1,1,1,1), nrow=4, byrow=TRUE)))
}

