% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigenereCipher.R
\name{letterCounts}
\alias{letterCounts}
\title{Get Letter Counts}
\usage{
letterCounts(txt)
}
\arguments{
\item{txt}{The text which you are analyzing}
}
\description{
Gets a count of the frequencies of each letter
}
\examples{
letterCounts("thisisan@examplesforlettercounts")
}

