% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigenereCipher.R
\name{letterFreq}
\alias{letterFreq}
\title{Find Relative Frequencies}
\usage{
letterFreq(txt)
}
\arguments{
\item{txt}{The text which you are analyzing}
}
\description{
Compute the relative frequencies of the letters in a string of lowercase letters.
}
\examples{
letterFreq("thisisan@examples")
}

