% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_tools.r
\name{md5Cipher}
\alias{md5Cipher}
\title{MD5 block cipher}
\usage{
md5Cipher(m, ky, iv)
}
\arguments{
\item{m}{A raw vector, representing the message to be encrypted/decrypted.
Assumed to have length a multiple of 16 (i.e., the message length is a multiple
of 128-bits).}

\item{ky}{A string, used as the symmetric key.}

\item{iv}{A string, used as an initial vector.}
}
\value{
A raw vector of the same length as \code{m}
}
\description{
This block cipher encrypts in 128-bit blocks. Encrypts a raw vector using a variation of
OFB mode. (Certainly not very secure.) The initial vector \code{iv} is converted to
a raw vector representing its MD5 hash. This raw vector becomes X_1. If the blocks of
\code{m} are P_1, P_2, ..., P_k, then O_i is the MD5 hash of X_i using \code{k} as the
HMAC key (as a string), where X_(i+1) = O_i. The ciphertext
block C_i is the exclusive-or of O_i and P_i.
}
\examples{
testct <- "This block cipher is for educational purposes only. Learn well!!"
ct <- md5Cipher(charToRaw(testct), "Runner4567", "init vector") # should end in: 78 be 24 73
rawToChar(md5Cipher(ct, "Runner4567", "init vector")) # decryption
}

