% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_tools.r
\name{miniSHA}
\alias{miniSHA}
\title{Miniature SHA hash}
\usage{
miniSHA(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
A four-character hash string.
}
\description{
Implements a 32-bit hash function by taking the first 32 bits of the SHA-1 hash.
}
\examples{
miniSHA("Corned Beef and Haggis") # should be "de0d"
}

