% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftcipher.R
\name{mod26ToString}
\alias{mod26ToString}
\title{Convert a vector of integers mod 26 to a string}
\usage{
mod26ToString(x)
}
\arguments{
\item{x}{A vector of integers (assumed to contain only 0...25)}
}
\value{
A string, where 'a' corresponds to 0, 'b' to 1, etc.
}
\description{
Convert a vector of integers mod 26 to a string
}
\examples{
mod26ToString(c(18,14,23,11,14,18,4))
}

