% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTPcipher.R
\name{oneTimePad}
\alias{oneTimePad}
\title{One-Time-Pad Cipher}
\usage{
oneTimePad(txt, padKey)
}
\arguments{
\item{txt}{The plaintext that will be encoded}

\item{padKey}{The key one time pad}
}
\value{
The encrypted text
}
\description{
Encrypt a message using a one time pad
}
\examples{
oneTimePad("test", "abcd")

}

