% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeTools2.R
\name{parityCheckMatrix}
\alias{parityCheckMatrix}
\title{Create a parity check matrix from a generator matrix}
\usage{
parityCheckMatrix(G)
}
\arguments{
\item{G}{A generator matrix of the form [I_k P]}
}
\value{
A parity check matrix of the form [P^T I_{n-k}].
}
\description{
Given a generator matrix for a systematic linear binary code,
return the parity check matrix.
}
\examples{
parityCheckMatrix(matrix(c(1,0,0,0,1,1,0,
                           0,1,0,0,1,0,1,
                           0,0,1,0,0,1,1,
                           0,0,0,1,1,1,1), nrow=4, byrow=TRUE))
}

