% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeTools3.R
\name{randBinVector}
\alias{randBinVector}
\title{` Generate a random binary vector}
\usage{
randBinVector(length, weight)
}
\arguments{
\item{length}{The number of components of the vector.}

\item{weight}{The Hamming weight of the vector. Assumed to be less than or equal to \code{length}.}
}
\value{
A numeric vector of 0's and 1's with specified length and weight.
}
\description{
Given a desired length and Hamming weight, return a binary vector (using
numeric variables) of the specified length and weight. Uses \code{sample} to
randomize.
}
\examples{
randBinVector(8,4)
}

