% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeTools3.R
\name{randInvMatrix}
\alias{randInvMatrix}
\title{Generate random binary matrix and inverse}
\usage{
randInvMatrix(n, numops = 100)
}
\arguments{
\item{n}{The dimension of the square matrix.}

\item{numops}{The number of random row operations to perform. Defaults to 100.}
}
\value{
A list, consisting of a random binary matrix M and its inverse.
}
\description{
Given a dimension n, this function uses a random sequence of elementary
row operations to produce an invertible n by n matrix, along with its
inverse. It does so by perfoming a sequence of \code{numops} randomly
chosen row operations on the identity matrix. It computes the inverse
at the same time by performing the inverse of these row operations in
reverse order.
}
\examples{
matrixList <- randInvMatrix(5)
matrixList$M
matrixList$Minv
}

