% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeTools3.R
\name{randPermMatrix}
\alias{randPermMatrix}
\title{Generate random permutation matrix}
\usage{
randPermMatrix(n, numswaps = 100)
}
\arguments{
\item{n}{The dimension of the matrix.}

\item{numswaps}{The number of random swaps to perform. Defaults to 100.}
}
\value{
A randomly generated n by n permutation matrix.
}
\description{
Given a dimension n, this function generates a random n by n permutation matrix. It
does so by performing a sequence of \code{numswaps} randomly chosen row swaps on the
identity matrix.
}
\examples{
randPermMatrix(5)
}

