% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftcipher.R
\name{shiftCipher}
\alias{shiftCipher}
\title{Encrypt a string using a shift cipher}
\usage{
shiftCipher(plainText, shift)
}
\arguments{
\item{plainText}{A string of lowercase letters}

\item{shift}{An integer (mod 26) to shift by}
}
\value{
An encrypted (or decrypted) string, where each letter has been shifted by \code{shift} places.
}
\description{
Encrypt a string using a shift cipher
}
\examples{
plainText <- "thisisasupersecretmessagethatwewanttoencrypt"
cipherText <- shiftCipher(plainText, 5)
print(cipherText)
print(shiftCipher(cipherText, 21)) # should be original plaintext
}

