% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigenereCipher.R
\name{shiftVec}
\alias{shiftVec}
\title{Shift Vector
 Shift a vector by n places}
\usage{
shiftVec(v, n)
}
\arguments{
\item{v}{The vec}

\item{n}{The shift}
}
\description{
Shift Vector
 Shift a vector by n places
}
\examples{
shiftVec(c(0,1,2,3), 5)
}

