% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigenereCipher.R
\name{skipString}
\alias{skipString}
\title{Skip String}
\usage{
skipString(txt, n, r)
}
\arguments{
\item{txt}{The string that is being shifted}

\item{n}{The shift}

\item{r}{The scale}
}
\description{
Given a string txt and integers n and r, returns a string consisting of the characters in positions n,n+r,n+2r,n+3r,…n,n+r,n+2r,n+3r,…, etc
}
\examples{
skipString("thisisan@examples", 4, 2)
}

