% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsaTools.r
\name{stringToBigz}
\alias{stringToBigz}
\title{Convert string to big integer}
\usage{
stringToBigz(txt)
}
\arguments{
\item{txt}{A character string}
}
\value{
A `bigz` integer
}
\description{
Converts a string to a bigz integer. First the characters of the string are converted to raw, then the
raw (hexadecimal) vector is converted to an integer, where the place values of this vector are assigned
from right to left.
}
\examples{
stringToBigz("A string, which may contain punctuation.")
}

