% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftcipher.R
\name{stringToMod26}
\alias{stringToMod26}
\title{Convert a string to a vector of integers mod 26}
\usage{
stringToMod26(x)
}
\arguments{
\item{x}{A string (assumed to contain only lowercase letters)}
}
\value{
A vector of integers, where 0 corresponds to 'a', 1 to 'b', etc.
}
\description{
Convert a string to a vector of integers mod 26
}
\examples{
stringToMod26("cubswin")
}

