% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeTools2.R
\name{syndrome}
\alias{syndrome}
\title{Syndrome of a vector}
\usage{
syndrome(r, H)
}
\arguments{
\item{r}{An n-dimensional numeric vector of 0's and 1's.}

\item{H}{A parity check matrix for an [n,k] code.}
}
\value{
A character string giving the syndrome of \code{r}.
}
\description{
Given an n-dimensional binary vector and a parity check matrix
for a binary linear [n,k] code, returns the syndrome of the vector,
as a character string of "1"'s and "0"'s.
}
\examples{
syndrome(c(1,1,0,1,0,1,1), matrix(c(1,1,0,1,1,0,0,
                                    1,0,1,1,0,1,0,
                                    0,1,1,1,0,0,1),
                                    nrow=3, byrow=TRUE)) # "111"
}

