% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigenereCipher.R
\name{vigenere}
\alias{vigenere}
\title{Vigenere Cipher}
\usage{
vigenere(txt, keyVector)
}
\arguments{
\item{txt}{The text which you are encrypting}

\item{keyVector}{A numerical vector used as the encryption key}
}
\description{
Encrypt a message using a vigenere cipher
}
\examples{
vigenere("thisistheplaintext", stringToMod26("thiswillbeavector"))
}

